+++ index.html
@@ -210,7 +210,7 @@
       </ul>
       <p class="popup__description">Великолепная квартира-студия в центре Токио. Подходит как туристам, так и бизнесменам. Квартира полностью укомплектована и недавно отремонтирована.</p>
       <div class="popup__photos">
-        <img src="" class="popup__photo" width="45" height="40" alt="Фотография жилья">
+        <img src="#" class="popup__photo" width="45" height="40" alt="Фотография жилья">
       </div>
     </article>
   </template>
@@ -229,7 +229,7 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script src="js/main.js"></script>
+  <script type="module" src="js/script.js"></script>
 
 </body>
